ref 🧼: - Cleaned up and made attendance code easier - Made the main lists sort nicely
